requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

oslo.config == 2.2.0
python-keystoneclient
docker-py == 1.3.1